fix(hopper_gemm): add unittest and small fix. (#49)

+++ Cuda/hopper_gemm/hopper_gemm/run_build.sh
@@ -2,6 +2,10 @@
 
 cd build
 
+if [ -f "hopper_gemm" ]; then
+    rm hopper_gemm
+fi
+
 if [ -f "CMakeCache.txt" ]; then
     rm CMakeCache.txt
 fi
@@ -14,6 +18,10 @@ cmake ..
 
 make 2>&1 | tee ../build.log
 
-./hopper_gemm 2>&1 | tee ../run.log
-
+if [ -f "hopper_gemm" ]; then
+    echo "build success"
+    ./hopper_gemm 2>&1 | tee ../run.log
+else
+    echo "build failed"
+fi
 cd ..
